+++ src/libaqualink/mqtt/mqtt_utils.c
@@ -12,16 +12,16 @@ char* generate_mqtt_id(char* buf, int len)
 {
 	strncpy(buf, AQUALINKD_NAME, len);
 	
-	int i = strlen(buf);
+	size_t i = strlen(buf);
 
 	if (i < len)
 	{
 		buf[i++] = '_';
 
 		// If we can't get MAC to pad mqtt id then use PID
-		if (!mac(&buf[i], len - i))
+		if (!mac(&buf[i], (int)(len - i)))
 		{
-			sprintf(&buf[i], "%.*d", (len - i), GetPid());
+			sprintf(&buf[i], "%.*d", (int)(len - i), GetPid());
 		}
 	}
